refactor: 항공권 도메인 코드 아키텍처를 헥사고날 아키텍처로 수정
