+++ backend/nest-backend/src/flight/flight.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { FlightService } from './flight.service';
-import { FlightController } from './adapter/in/flight.controller';
+import { FlightService } from './application/service/flight.service';
+import { FlightController } from './adapter/in/web/flight.controller';
 
 @Module({
   controllers: [FlightController],
